review adjustment
