+++ src/apimachinery/apiserver/field_template/field_template.go
@@ -49,3 +49,28 @@ func (t template) ListFieldTemplate(ctx context.Context, h http.Header, opt *met
 
 	return &resp.Data, nil
 }
+
+// ListObjFieldTmplRel list field template and object relations.
+func (t template) ListObjFieldTmplRel(ctx context.Context, h http.Header, opt *metadata.ListObjFieldTmplRelOption) (
+	*metadata.ObjFieldTmplRelInfo, errors.CCErrorCoder) {
+
+	resp := new(metadata.ListObjFieldTmplRelResp)
+
+	err := t.client.Post().
+		WithContext(ctx).
+		Body(opt).
+		SubResourcef("/findmany/field_template/object/relation").
+		WithHeaders(h).
+		Do().
+		Into(resp)
+
+	if err != nil {
+		return nil, errors.CCHttpError
+	}
+
+	if err := resp.CCError(); err != nil {
+		return nil, err
+	}
+
+	return &resp.Data, nil
+}
